+++ implementations/networks/torch/policy/atari.py
@@ -190,8 +190,8 @@ class Atari_Core(Policy_Network, nn.Module, Safe_nn_Module):
         logits_position = self.position_step(torch.concat([last_position, action_onehot], dim=-1))
         props_position = Bernoulli(probs=logits_position)
 
-        log_prob_flag = props_flag.log_prob(action_flag.long())
-        log_prob_action = props_action.log_prob(action_action.long())
+        log_prob_flag = props_flag.log_prob(action_flag)
+        log_prob_action = props_action.log_prob(action_action)
         log_prob_position = props_position.log_prob(current_position).mean(-1)
         log_prob_content = props_content.log_prob(action_content).mean(-1)
         
@@ -237,9 +237,9 @@ class Atari_Core(Policy_Network, nn.Module, Safe_nn_Module):
             raise ValueError("At least one of b_int, b_ext, b_content must be provided")
         
         if b_int is None:
-            b_int = np.zeros((batch_size,), dtype=int)
+            b_int = np.zeros((batch_size,), dtype=float)
         if b_ext is None:
-            b_ext = np.zeros((batch_size, 1), dtype=int)
+            b_ext = np.zeros((batch_size, 1), dtype=float)
         if b_position is None:
             b_position = np.zeros((batch_size, self.position_size), dtype=float)
         if b_content is None:
